Renamed context to providers
