+++ src/components/board/board-deck-dialog.tsx
@@ -10,7 +10,7 @@ import {
   DialogTrigger,
 } from '@/components/ui/dialog';
 import { FC } from 'react';
-import { useBoardContext } from '../../context/board-context';
+import { useBoardContext } from '../../providers/board-provider';
 
 const BoardDeckDialog: FC = () => {
   const { deckCount, players, onDrawCard, gameContext } = useBoardContext();
