+++ src/components/ui/dialog.tsx
@@ -5,7 +5,7 @@ import { X } from 'lucide-react';
 import * as React from 'react';
 
 import { cn } from '@/lib/utils';
-import { useFullScreen } from '../../context/fullscreen-context';
+import { useFullScreen } from '../../providers/fullscreen-provider';
 
 const Dialog = DialogPrimitive.Root;
