+++ src/components/ui/popover.tsx
@@ -4,7 +4,7 @@ import * as PopoverPrimitive from '@radix-ui/react-popover';
 import * as React from 'react';
 
 import { cn } from '@/lib/utils';
-import { useFullScreen } from '../../context/fullscreen-context';
+import { useFullScreen } from '../../providers/fullscreen-provider';
 
 const Popover = PopoverPrimitive.Root;
